+++ app/src/main/java/atreides/house/arxiv_r/FirstFragment.java
@@ -8,7 +8,6 @@ import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
 import android.util.Log;
 import android.view.LayoutInflater;
-import android.view.MotionEvent;
 import android.view.View;
 import android.view.ViewGroup;
 import java.util.List;
@@ -20,11 +19,11 @@ import java.util.List;
 public class FirstFragment extends Fragment {
     View myView;
 
+    public RecyclerView.Adapter mAdapter;
     private static final int REQUEST_EXTERNAL_STORAGE = 1;
     private static String[] PERMISSIONS_STORAGE = {
             Manifest.permission.READ_EXTERNAL_STORAGE,
             Manifest.permission.WRITE_EXTERNAL_STORAGE};
-    private RecyclerView mRecyclerView;
 
     @Override
     public void onCreate(Bundle savedInstanceState) {
@@ -42,9 +41,14 @@ public class FirstFragment extends Fragment {
         List<RssFeedModel> mFeedModelList = pal.getThing();
 
         myView = inflater.inflate(R.layout.first_layout, container, false);
-        mRecyclerView = myView.findViewById(R.id.recyclerView);
+        RecyclerView mRecyclerView = myView.findViewById(R.id.recyclerView);
         mRecyclerView.setLayoutManager(new LinearLayoutManager(getActivity()));
-        mRecyclerView.setAdapter(new RssFeedListAdapter(mFeedModelList));
+        mAdapter = new RssFeedListAdapter(mFeedModelList);
+        mRecyclerView.setAdapter(mAdapter);
         return myView;
+    }
+    public void newContent() {
+        // placeholder for if the user scrolls to the bottom like a noob
+
     }
 }
