+++ app/src/main/java/atreides/house/arxiv_r/MainActivity.java
@@ -179,11 +179,14 @@ public class MainActivity extends AppCompatActivity
         if (id == R.id.nav_first_layout) {
             category = "cs";
             setTitle("Computer Science");
-            checkExists();
+            new FetchFeedTask().FetchFeedTask(category,true);
         } else if (id == R.id.nav_second_layout) {
             category = "math";
-            checkExists();
             setTitle("Mathematics");
+            new FetchFeedTask().FetchFeedTask(category,true);
+
+            // ahhhhhhhh
+            String urlLink = ("https://export.arxiv.org/api/query?search_query=cat:" + category + "*&max_results=5");
         } else if (id == R.id.nav_slideshow) {
 
         } else if (id == R.id.nav_manage) {
@@ -199,232 +202,6 @@ public class MainActivity extends AppCompatActivity
         return true;
     }
 
-    public List<RssFeedModel> parseFeed(InputStream inputStream) throws XmlPullParserException, IOException {
-        String title = null;
-        String summary = null;
-        String author = null;
-        String published = null;
-        String updated = null;
-        String id = null;
-        boolean isItem = false;
-        List<RssFeedModel> items = new ArrayList<>();
-
-        try {
-            XmlPullParser xmlPullParser = Xml.newPullParser();
-            xmlPullParser.setFeature(XmlPullParser.FEATURE_PROCESS_NAMESPACES, false);
-            xmlPullParser.setInput(inputStream, null);
-
-            xmlPullParser.nextTag();
-            while (xmlPullParser.next() != XmlPullParser.END_DOCUMENT) {
-                int eventType = xmlPullParser.getEventType();
-
-                String name = xmlPullParser.getName();
-                if (name == null)
-                    continue;
-
-                if (eventType == XmlPullParser.END_TAG) {
-                    if (name.equalsIgnoreCase("entry")) {
-                        isItem = false;
-                    }
-                    continue;
-                }
-
-                if (eventType == XmlPullParser.START_TAG) {
-                    if (name.equalsIgnoreCase("entry")) {
-                        isItem = true;
-                        continue;
-                    }
-                }
-
-                Log.d("MainActivity", "Parsing name ==> " + name);
-                String result = "";
-                if (xmlPullParser.next() == XmlPullParser.TEXT) {
-                    result = xmlPullParser.getText();
-                    xmlPullParser.nextTag();
-                }
-
-                if (name.equalsIgnoreCase("title")) {
-                    title = result;
-                } else if (name.equalsIgnoreCase("summary")) {
-                    summary = result;
-                } else if (name.equalsIgnoreCase("author")) {
-                    xmlPullParser.next();
-                    author = xmlPullParser.getText();
-                } else if (name.equalsIgnoreCase("published")) {
-                    published = result;
-                } else if (name.equalsIgnoreCase("updated")) {
-                    updated = result;
-                } else if (name.equalsIgnoreCase("id")) {
-                    id = result;
-                }
-                Log.d("MainActivity", "title = " + title + "\nsummary = " + summary + "\nauthor = " + author);
-                if (title != null && summary != null && author != null) {
-                    if (isItem) {
-                        RssFeedModel item = new RssFeedModel(title, summary, author, published, updated, id);
-                        items.add(item);
-                        Log.d("MainActivity", "item added");
-                    } else {
-                        mFeedTitle = title;
-                        mFeedSummary = summary;
-                        mFeedAuthor = author;
-                        mFeedPublished = published;
-                        mFeedUpdated = updated;
-                        mFeedId = id;
-                    }
-                    title = null;
-                    summary = null;
-                    author = null;
-                    published = null;
-                    updated = null;
-                    isItem = false;
-                }
-            }
-
-            return items;
-        } finally {
-            inputStream.close();
-        }
-    }
-
-    private class FetchFeedTask extends AsyncTask<Void, Void, Boolean> {
-        private String urlLink;
-        Dialog loadingBlip = new Dialog(MainActivity.this);
-
-        @Override
-        protected void onPreExecute() {
-            // pls wait for load my dude
-            loadingBlip.requestWindowFeature(Window.FEATURE_NO_TITLE);
-            loadingBlip.setContentView(R.layout.loading_blip);
-            loadingBlip.getWindow().setBackgroundDrawable(new ColorDrawable(Color.TRANSPARENT));
-
-            loadingBlip.show();
-
-            mFeedTitle = "title";
-            mFeedSummary = "summary";
-            mFeedAuthor = "author";
-            mFeedPublished = "published";
-            mFeedUpdated = "updated";
-            mFeedId = "id";
-
-            Log.d("MainActivity", "Category (when it counts) = " + category);
-            urlLink = ("https://export.arxiv.org/api/query?search_query=cat:" + category + "*&max_results=5");
-        }
-
-        @Override
-        protected Boolean doInBackground(Void... voids) {
-
-            try {
-                URL url = new URL(urlLink);
-                HttpsURLConnection conn = (HttpsURLConnection) url.openConnection();
-                InputStream inputStream = conn.getInputStream();
-                mFeedModelList = parseFeed(inputStream);
-                return true;
-
-            } catch (IOException e) {
-                Log.e(TAG, "Error", e);
-            } catch (XmlPullParserException e) {
-                Log.e(TAG, "Error", e);
-            }
-            return false;
-        }
-
-        @Override
-        protected void onPostExecute(Boolean success) {
-            Log.d("postexecute","has begun");
-            if (success) {
-                loadingBlip.dismiss();
-                sendMessenger();
-            } else {
-                Toast.makeText(MainActivity.this,
-                        "Questionable internet connectivity...",
-                        Toast.LENGTH_LONG).show();
-            }
-        }
-    }
-
-    private void sendMessenger() {
-
-        try {
-            FileOutputStream fos = this.openFileOutput(category + "File", MODE_PRIVATE);
-            ObjectOutputStream of = new ObjectOutputStream(fos);
-            of.writeObject(mFeedModelList);
-            of.flush();
-            of.close();
-            fos.close();
-            Log.d("sendMessenger","Closed");
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-        String test = "Rubber duckies are the SHIT";
-        FragmentManager fragmentManager = getFragmentManager();
-        FirstFragment newFragment = new FirstFragment();
-        ParcelableArrayList pal = new ParcelableArrayList();
-        pal.setThing(mFeedModelList);
-        pal.setTitle(test);
-        Log.d("sendMessenger", String.valueOf(mFeedModelList));
-        Bundle bundle = new Bundle();
-        bundle.putParcelable("articles", pal);
-        bundle.putString("test", test);
-        if (bundle != null) {
-            newFragment.setArguments(bundle);
-            fragmentManager.beginTransaction()
-                    .replace(R.id.content_frame
-                            , newFragment)
-                    .addToBackStack(null)
-                    .commit();
-        } else {
-            Log.d("MainActivity", "It's null");
-        }
-    }
-
-    private void checkExists() {
-        // see if file exists
-        File catFile = getApplicationContext().getFileStreamPath(category + "File");
-        if (catFile == null || !catFile.exists()) {
-            Log.d("redundancy checker","doesn't exist");
-            // This means we should write a new one
-            new FetchFeedTask().execute((Void) null);
-        } else {
-            Log.d("redundancy checker","exists");
-
-            // file exists, so check date
-            SimpleDateFormat df = new SimpleDateFormat("MMdd");
-            Date todayDate = new Date();
-            Date lastModDate = new Date(catFile.lastModified());
-            String today = df.format(todayDate);
-            String lastMod = df.format(lastModDate);
-
-            if (Integer.parseInt(today) != Integer.parseInt(lastMod)) {
-                // outdated, get new
-                Log.d("wat",today +" is today and last mod is " + lastMod);
-                Log.d("redundancy checker","outdated, fetching new data");
-                new FetchFeedTask().execute((Void) null);
-
-            } else {
-                // current, send cached
-                Log.d("redundancy checker","current");
-
-                ArrayList<RssFeedModel> cachedData;
-                FileInputStream fis;
-                try {
-                    fis = this.openFileInput(category + "File");
-                    ObjectInputStream oi = new ObjectInputStream(fis);
-                    cachedData = (ArrayList<RssFeedModel>) oi.readObject();
-                    oi.close();
-                    fis.close(); // hmmmm
-                    mFeedModelList = cachedData;
-                } catch (FileNotFoundException e) {
-                    Log.e("InternalStorage", e.getMessage());
-                } catch (IOException e) {
-                    Log.e("InternalStorage", e.getMessage());
-                } catch (ClassNotFoundException e) {
-                    e.printStackTrace();
-                }
-                Log.d("redundancy checker","sending old info");
-                sendMessenger();
-            }
-        }
-    }
     @Override
     public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
         super.onRequestPermissionsResult(requestCode, permissions, grantResults);
